% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net}
\alias{net}
\title{Calculate net}
\usage{
net(journ, classselect = "all", accselect = "all", dates = "all")
}
\arguments{
\item{journ}{Journal dataframe}

\item{classselect}{Which classes to include. c("class1","class2",...). Defaults to "all".}

\item{accselect}{Which accounts to include. c("class1","class2",...). Defaults to "all".}

\item{dates}{Date range to include. c("2014-01-01","2014-12-31"). Defaults to "all".}
}
\description{
This function reads in your journal, subsets it according to queries, and calculates net.
}
\examples{
net()
}
\keyword{net}

